Move a cast in Buffer.indexOf().
This has the happy side-effect of preventing the Lenovo S939 from
behaving incorrectly when executing this method.

The following test exercises the previous code; it fails after
a few iterations on that device:
https://gist.github.com/swankjesse/4396e276e43dbb0e71cc

Unfortunately there's no good way to regression test that we
haven't hit a bug in a specific lousy device.