Make Source.indexOf(ByteString) significantly faster.
Previously the algorithm that did this was extremely inefficient, and had worst
case runtime of O(N * S * S) for N is size of the bytestring and S is the number
of segments.

The new code runs in O(N * S). It accomplishes this by not starting each
search at the first segment, which could occur many times when called by
RealBufferedSource.