Fix a major bug where HashingSource is incorrect for large writes.
There was a bug where it wasn't traversing through the segments of
the buffer being hashed. This means that HashingSource was returning
incorrect answers for any writes that spanned multiple segment
boundaries.

The problem wasn't detected by our test cases because although
we have tests for non-uniform segments, and for buffers that
span multiple segments, we don't have tests for non-uniform
buffers that span multiple segments. This has been fixed too.