Allow Buffer.getByte() to search backwards for the desired segment
Uses of Buffer.getByte() withing the Okio project access it more often
relative to the end of the Buffer instead of from the beginning.
Specifically the ReadBufferedSource.readHexadecimalUnsignedLong() and
ReadBufferedSource.readDecimalLong() methods make heavy use of
retrieving the latest byte from the Buffer and with this change, the
unit tests for these methods with the OneByteAtATime source no longer
take over a second to run.

A benchmark for this behavior has been created and verified to show
significant improvement.