Samples which show integration with Java NIO channels
First part of the samples shows how to create a Source or Sink from a
ReadableByteChannel or a WritableByteChannel respectively. This uses the
UnsafeCursor to give the NIO channel direct access to the bytes within
the Buffer by wrapping the internal segments with a ByteBuffer.

The second part shows how to take advantage of the FileChannel's
transferFrom and transferTo methods. Again, this allows direct access to
the data being read or written, but this time by giving the Buffer the
ByteBuffers used by the FileChannel.