Don't use hash function names in non-JVM platforms
Also replace HashFunctionTest.kt with MessageDigestConsistencyTest.kt, which
exercises many more access patterns.

This creates a directory of .kt files, okio/src/hashFunctions/kotlin/ that
is included by both nonJvmMain and jvmTest. I would have preferred to create
a sourceSet for hashFunctions, but that was causing Kotlin Multiplatform
builds to fail.

This doesn't restore the hash functions to the benchmark, though that should
be possible; either by depending on the test artifact or by creating a JVM
only artifact that is depended-upon by both test and JMH.