FileSystem.atomicMove() (#806)
* FileSystem.atomicMove()

I didn't implement the full move operation that works across filesystems.
That would be some straightforward code to copy a source file to a target
file using our existing source() and sink() abstractions. But it would also
require some not-straightforward code to copy other metadata such as file
permissions, created and modified dates, and extended attributes. That
method has awkward behavior in partial failures; the source file exists
and the target file does too, but possibly partially.

In real application code my preferred way of doing this would be a
copy to a temporary file on the target filesystem followed by an
atomic move of the temp file on that filesystem. This way if the
move fails partially the target file doesn't exist. Unfortunately
that strategy isn't appropriate for Okio because it potentially
leaves a temporary file behind with no mechanism to clean it up.

Punting to the developer is okay for now; we may want to offer
a userspace  sample or built-in function that
does some or all of the above.

* Update okio-files/src/jvmMain/kotlin/okio/JvmSystemFilesystem.kt

Co-authored-by: Jake Wharton <github@jakewharton.com>

* Update okio-files/src/jvmMain/kotlin/okio/JvmSystemFilesystem.kt

Co-authored-by: Jake Wharton <github@jakewharton.com>

Co-authored-by: Jake Wharton <github@jakewharton.com>