Prevent calling scheduleTimeout twice on an AsyncTimeout
Invoking `enter` concurrently on an `AsyncTimeout` instance may lead to
`scheduleTimeout` being invoked twice.  If `scheduleTimeout` is called
on an `AsyncTimeout` instance that is already scheduled, it could lead
to a cycle in the single linked list starting with `head`, which in turn
would lead to an infinite loop in `cancelScheduledTimeout`.