Run tests on Android
This is currently completely manual. In a follow-up it would be nice
to get this running in CI.

On a Pixel 3A emulator running Android 30 I observed test crashes in
the following tests:

    selectSpanningMultipleSegments
    readLargeNioBufferOnlyReadsOneSegment
    gzipSink
    gzipSource

I needed to rename BufferedSourceTest to BufferedSourceJavaTest, and similarly
for BufferedSinkJavaTest and ByteStringJavaTest. We were seeing simple name
collisions in these tests which I believe were previously preventing these tests
from being executed.