Remove the slash property from Path
This was an attempt to make Paths have a property that would
differentiate whether they follow Windows vs. UNIX rules. But
the consequence as that two relative paths, 'foo', and 'foo'
could be different because of the hidden state of the unused
slash.

Instead just look inside the path for the slash. In the case
where we're resolving two paths without either having any
slash characters internally, just use the platform slash.