Fix volume letter parsing in Windows Paths that contain forward slashes.
This PR removes the requirement for a backslash to be present for volume letters to be parsed.

Note that this doesn't fix the case of UNC paths that use forward slashes. A test like the following still fails:

```kotlin
@Test  fun windowsUncAbsolutePathForwardSlashes() {
  val path = "//server/project/notes.txt".toPath()
  assertEquals("//server/project/notes.txt", path.toString())
  assertEquals("//server/project".toPath(), path.parent)
  assertNull(path.volumeLetter)
  assertEquals("notes.txt", path.name)
  assertTrue(path.isAbsolute)
  assertFalse(path.isRoot)
}
```

Since `//foo` is valid as both a UNC path and a POSIX path, I don't believe paths like that can be parsed unambiguously without the notion of a system file separator or similar.

[1]: https://docs.microsoft.com/en-us/windows/win32/api/fileapi/nf-fileapi-createfilea#parameters
[2]: https://docs.microsoft.com/en-us/dotnet/standard/io/file-path-formats#canonicalize-separators