FileHandle, for reading and writing, streaming and random access (#916)
I'm hoping to get something close to the underlying file descriptor or file
handle abstraction in POSIX.

I am unsatisfied with the design of okio.Cursor and how it will work with
random-access reading and writing. I'm hoping to make this the successor to
that class, and to completely delete that class from our public API.

I am planning a follow up PR on FileSystem, FileSystem.open(Path) that
returns a FileHandle. I will also make FileSystem.source() and
FileSystem.sink() all delegate to FileSystem.open().