Provide even more information when FileSystem timestamps surprise
I observed this failure in CI:
>  expected 2021-05-29T22:15:58.995Z in 2021-05-29T22:15:59Z..2021-05-29T22:16:01Z

I can't tell if the problem is that the file system is using a different clock, or
if the time bounds are being rounded too much.

My best guess is that the CI filesystem is using a different clock from the test
and there's at least 5 milliseconds of clock skew between them.