APIs to seek a FileHandle's source or sink (#959)
The one fancy thing we're doing here is preserving the buffered data if
we're seeking forward on a source.

I don't think there's a compelling equivalent for preserving buffered
data when seeking backwards on a sink. In theory we could truncate the
buffer to the new seek position, but this breaks if the writer doesn't
overwrite the full range of what was truncated.