Restore support for JS browser platforms
When I introduced the Node FileSystem I broke support for running Okio with
JavaScript in the browser. This restores that break.

This change introduces a new okio-testing-support module so that the
AbstractFileSystemTest can be shared by both okio and okio-nodefilesystem
modules. I had to move NodeJsFileSystemTest to the okio-nodefilesystem
module because the core okio module no longer depends on NPM, not even in
testing.