Move NSData to ByteString conversion onto ByteString companion
This is a convenience for Objeective-C/Swift callers such that when a ByteString is exposed in your header you are also given this (and any future) conversion functions as automatically available.

Because SegmentedByteString relies on ByteString's internal constructor which is only avaialble in the actuals it must be duplicated in the two source sets.