Implement `FileHandle` for asset file system (#1254)
* Implement FileHandle for asset file system

While uncompressed assets can be accessed using a file descriptor and normal random access methods, most assets will be compressed and thus can only be read linearly. As such, we maintain a single stream that only goes forward. If we need to go
backward for some reason, the old stream is closed and a new one is created.

* Update okio-assetfilesystem/src/main/kotlin/okio/assetfilesystem/AssetFileSystem.kt

Co-authored-by: Jesse Wilson <jwilson@squareup.com>

---------

Co-authored-by: Jesse Wilson <jwilson@squareup.com>