New API, BufferedSource.indexOf(ByteString, fromIndex, toIndex)
This is surprisingly interesting. To minimize unnecessary reads for
toIndex it is necessary to check whether a prefix of the query matches
a suffix of the currently-loaded data.

This read-avoidance is useful in practice. When doing HTTP multipart
decoding the caller may scan for a boundary separator with a bounded
range, and we don't want to block reading when doing so won't impact
the result of the call.