Re-implement Buffer.rangeEquals on Buffer.indexOf (#1628)
The old rangeEquals function was embarassingly inefficient,
using a full scan of the segments linked-list for each element.
That optimization is already done in indexOf(), which has
almost all of what we need for rangeEquals().

This adds an offset and count slice to the input ByteString
in indexOf(), and then leverages that to simplify rangeEquals().

Co-authored-by: Jesse Wilson <jwilson@squareup.com>