New concrete implementation of okio.Socket
This turned out to be much trickier than I had expected.
The default java.net.Socket inputStream and outputStream
close the entire Socket when they are closed. This offers
different behavior - the two streams are closed independently.
When both are closed the socket is closed and its resources
are released.