Get Socket.inputStream and outputStream eagerly (#1669)
I was looking into migrating OkHttp to use Okio's
Socket internally and I came across a small behavior
change - previously we were getting the stream once
and reusing it on each call, but now Okio is getting
the stream on each access.

With this change the Socket.asOkioSocket() API behaves
more consistently with Socket.source() and Socket.sink().

Co-authored-by: Jesse Wilson <jwilson@squareup.com>